video: Fix pointer size
Pointer could always hold only 32 bit value which is not enough on 64 bit
platform.

+++ include/video_fb.h
@@ -38,7 +38,7 @@ typedef struct graphic_device {
     unsigned int dprBase;
     unsigned int vprBase;
     unsigned int cprBase;
-    unsigned int frameAdrs;
+    unsigned long frameAdrs;
     unsigned int memSize;
     unsigned int mode;
     unsigned int gdfIndex;
